requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
colorthief==0.2.1
python-dotenv==1.0.0
lxml==4.9.3
